add number of bed and type in HOuse add&Modify

+++ database/migrations/2017_07_11_175929_create_House_table.php
@@ -45,7 +45,7 @@ class CreateHouseTable extends Migration
 
                 $table->integer('size')->nullable();
                 $table->integer('numOfRooms')->nullable();
-                $table->integer('numOfBaths')->nullable();
+                $table->float('numOfBaths')->nullable();
                 $table->integer('numOfBeds')->nullable();
                 $table->integer('maxNumOfGuests')->nullable();
